Fixes issue 113: Allows for proper reading of nested private sequences
In implicit VR DICOM files, the VR is determined based upon the tag dictionary. If a private tag is encountered, the VR is undefined. As a result, the data is read in as a byte string. This is problematic if the private element is a sequence.

If a sequence item tag is found immediately after the length information, then the VR can be assumed to be 'SQ'. In this way, nested private sequences can be parsed properly and the remainder of the file can be read in.

See issue 113 for more information.

This commit provides a simple example DICOM file as well as the appropriate unittests.

All unittests pass for Python 2.6-3.2.

Initial fix provided by David Hunter