Prevents charset.decode from attempting to decode unicode values
Unicode strings are already decoded, and thus can't be decoded again. Previously, if Dataset.decode was called twice on the same Dataset, an error would result.

Additionally, this commit moves the conversion of DICOM encodings into python encodings into a standalone function to allow other parts of the module to grab the appropriate encoding values.