py3: Fixes python3 incompatibility with using __slots__ for classes derived from int
Currently, IS is a subclass of the built-in int datatype. Because of the C implementation of ints in python3, __slots__ can no longer be used for optimization in classes derived from int.

An alternate version of this commit would be to simply remove the use of __slots__ in just the IS class definition.

The logic in this commit does not change the performance significantly.